Modified initialize_benchmark function to consider existence of benchmark.csv

+++ scripts/benchmark.py
@@ -1,31 +1,32 @@
 import os
 import csv
 
-BENCHMARK_FILE = 'benchmark.csv'
+BENCHMARK_DIRECTORY = os.path.join(os.path.abspath(os.pardir), "benchmark")
+BENCHMARK_FILE = os.path.join(BENCHMARK_DIRECTORY, "benchmark.csv") 
 
-
-def add_report(report):
-    benchmark_directory = os.path.join(os.path.abspath(os.pardir), "benchmark")
-    file_path = os.path.join(benchmark_directory, BENCHMARK_FILE) 
+# Gets a report dictionary and file path to a csv file
+# Inserts a row containing the report
+def add_report(report, file_path=BENCHMARK_FILE):
     row = [report['Embedding Model'],
            report['DB Type'],
            report['Strategy'],
            report['Average k'],
            report['Sigma']]
-    with open(file=file_path, mode='a') as fn:
+    with open(file=BENCHMARK_FILE, mode='a') as fn:
         csv_writter = csv.writer(fn, delimiter=' ', lineterminator='\n')
         csv_writter.writerow(row)
 
-def initialize_benchmark():
-    benchmark_directory = os.path.join(os.path.abspath(os.pardir), "benchmark")
-    file_path = os.path.join(benchmark_directory, BENCHMARK_FILE) 
+#   Initializes the benchmark file if not available
+def initialize_benchmark(file_path=BENCHMARK_FILE):
+    if os.path.exists(file_path):
+        return
     row = ['Embedding Model',
            'DB Type',
            'Strategy',
            'Average k',
            'Sigma']
     with open(file=file_path, mode='w') as fn:
-        csv_writter = csv.writer(fn, delimiter=' ', lineterminator='\n')
+        csv_writter = csv.writer(fn, delimiter=',', lineterminator='\n')
         csv_writter.writerow(row)
 
 def main():
